requirements.txt
pyTelegramBotAPI==4.0.0
python-dotenv==0.21.0
